refactor watcher_unix
