+++ CMakeLists.txt
@@ -252,9 +252,9 @@ if(WIN32)
     add_executable(lemon ${GUI_TYPE} ${LEMON_FULL_SOURCES})
 else()
     if(APPLE)
-        add_executable(watcher_unix unix/watcher_macos.mm)
+        add_executable(watcher_unix unix/watcher_unix.cpp unix/watcher_macos.mm)
     else()
-        add_executable(watcher_unix unix/watcher_linux.cpp)
+        add_executable(watcher_unix unix/watcher_unix.cpp unix/watcher_linux.cpp)
     endif()
     configure_file(unix/watcher.qrc ${CMAKE_BINARY_DIR} COPYONLY)
     list(APPEND LEMON_FULL_SOURCES ${CMAKE_BINARY_DIR}/watcher.qrc)
