+++ unix/watcher_macos.mm
@@ -1,6 +1,6 @@
 /*
  * SPDX-FileCopyrightText: 2011-2019 Project Lemon, Zhipeng Jia
- * SPDX-FileCopyrightText: 2019-2022 Project LemonLime
+ * SPDX-FileCopyrightText: 2019-2023 Project LemonLime
  *
  * SPDX-License-Identifier: GPL-3.0-or-later
  *
@@ -25,8 +25,6 @@
 #include <sys/wait.h>
 #include <unistd.h>
 #include <string>
-
-int pid;
 
 template <typename header> static ssize_t calculateStaticMemoryUsage(int fd, int start_offset = 0) {
 	ssize_t res = 0;
@@ -95,36 +93,26 @@ template <typename header> static ssize_t calculateStaticMemoryUsage(int fd, int
 	return res;
 }
 
-void cleanUp(int dummy) {
-	kill(pid, SIGKILL);
-	exit(0);
-}
-
-std::string getCpuBrandString() {
+static std::string getCpuBrandString() {
 	char buf[1024];
 	size_t buflen = 1024;
 	sysctlbyname("machdep.cpu.brand_string", &buf, &buflen, NULL, 0);
 	return std::string(buf, buflen);
 }
 
-int main(int argc, char *argv[]) {
-	int isAppleSilicon = getCpuBrandString().find("Apple") != std::string::npos;
+int isAppleSilicon;
 
-	int timeLimit;
-	ssize_t memoryLimit;
-	sscanf(argv[5], "%d", &timeLimit);
-	timeLimit = (timeLimit - 1) / 1000 + 1;
-	sscanf(argv[6], "%zd", &memoryLimit);
+void initWatcher() {
+	isAppleSilicon = getCpuBrandString().find("Apple") != std::string::npos
+}
 
-	/* check static memory usage */
-	std::string fileName(argv[1]);
-	fileName = fileName.substr(1);
-	fileName = fileName.substr(0, fileName.find("\""));
+ssize_t calculateStaticMemoryUsage(const std::string& fileName)
+{
 	uint32_t magic;
 	ssize_t staticMemoryUsage = 0;
 	int fd = open(fileName.c_str(), O_RDONLY);
 	if (fd < 0) {
-		return 1;
+		return -1;
 	}
 
 	int rc =
@@ -141,10 +129,10 @@ int main(int argc, char *argv[]) {
 	if (rc != 0) {
 		// not a fat file, consider a mach-o file
 		if (lseek(fd, 0, SEEK_SET) < 0) {
-			return 1;
+			return -1;
 		}
 		if (read(fd, &magic, sizeof(magic)) != sizeof(magic)) {
-			return 1;
+			return -1;
 		}
 		if (magic == MH_MAGIC) {
 			staticMemoryUsage = calculateStaticMemoryUsage<mach_header>(fd);
@@ -156,84 +144,13 @@ int main(int argc, char *argv[]) {
 	}
 	close(fd);
 
-	if (staticMemoryUsage == -1) {
-		return 1;
-	}
-	if (staticMemoryUsage > memoryLimit * 1024 * 1024) {
-		printf("0\n%zd\n", staticMemoryUsage);
-		return 0;
-	}
+	return staticMemoryUsage;
+}
 
-	memoryLimit *= 1024 * (isAppleSilicon ? 4 : 1);
+ssize_t getMemoryRLimit(ssize_t memoryLimitInMB) {
+	return memoryLimitInMB * 1024 * (isAppleSilicon ? 4 : 1);
+}
 
-	pid = fork();
-
-	if (pid > 0) {
-		signal(SIGINT, cleanUp);
-		signal(SIGABRT, cleanUp);
-		signal(SIGTERM, cleanUp);
-		struct rusage usage;
-		int status;
-
-		if (wait4(pid, &status, 0, &usage) == -1)
-			return 1;
-
-		if (WIFEXITED(status)) {
-			if (WEXITSTATUS(status) == 1)
-				return 1;
-
-			printf("%d\n", (int)(usage.ru_utime.tv_sec * 1000 + usage.ru_utime.tv_usec / 1000));
-			printf("%d\n", (int)(usage.ru_maxrss) / (isAppleSilicon ? 4 : 1));
-
-			if (WEXITSTATUS(status) != 0)
-				return 2;
-
-			return 0;
-		}
-
-		if (WIFSIGNALED(status)) {
-			printf("%d\n", (int)(usage.ru_utime.tv_sec * 1000 + usage.ru_utime.tv_usec / 1000));
-			printf("%d\n", (int)(usage.ru_maxrss) / (isAppleSilicon ? 4 : 1));
-
-			if (WTERMSIG(status) == SIGXCPU)
-				return 3;
-
-			if (WTERMSIG(status) == SIGKILL)
-				return 4;
-
-			if (WTERMSIG(status) == SIGABRT)
-				return 4;
-
-			return 2;
-		}
-	} else {
-		if (strlen(argv[2]) > 0)
-			freopen(argv[2], "r", stdin);
-
-		if (strlen(argv[3]) > 0)
-			freopen(argv[3], "w", stdout);
-
-		if (strlen(argv[4]) > 0)
-			freopen(argv[4], "w", stderr);
-
-		rlimit memlim, stalim, timlim;
-
-		if (memoryLimit > 0) {
-			memlim = (rlimit){(rlim_t)memoryLimit, (rlim_t)memoryLimit};
-			stalim = (rlimit){(rlim_t)memoryLimit, (rlim_t)memoryLimit};
-		} else {
-			memlim = (rlimit){RLIM_INFINITY, RLIM_INFINITY};
-			stalim = (rlimit){(rlim_t)2147483647ll, (rlim_t)2147483647ll};
-		}
-
-		timlim = (rlimit){(rlim_t)timeLimit, (rlim_t)(timeLimit + 1)};
-		setrlimit(RLIMIT_AS, &memlim);
-		setrlimit(RLIMIT_STACK, &stalim);
-		setrlimit(RLIMIT_CPU, &timlim);
-
-		if (execlp("bash", "bash", "-c", argv[1], NULL) == -1)
-			return 1;
-	}
-
-	return 0;
+size_t getMaxRSSInByte(long ru_maxrss) {
+	return ru_maxrss / (isAppleSilicon ? 4 : 1)
 }
